+++ apps/web/components/user/profile-tabs.tsx
@@ -1,0 +1,480 @@
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@workspace/ui/components/select';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@workspace/ui/components/tabs';
+import { Heart, TrendingUp, GitFork, Clock, ExternalLink } from 'lucide-react';
+import ProjectCard from '@/app/(public)/(projects)/projects/project-card';
+import { Card, CardContent } from '@workspace/ui/components/card';
+import { Skeleton } from '@workspace/ui/components/skeleton';
+import { Button } from '@workspace/ui/components/button';
+import { ContributionGraph } from './contribution-graph';
+import { Badge } from '@workspace/ui/components/badge';
+import Icons from '@workspace/ui/components/icons';
+import Link from '@workspace/ui/components/link';
+import { useQuery } from '@tanstack/react-query';
+import { cn } from '@workspace/ui/lib/utils';
+import { useTRPC } from '@/hooks/use-trpc';
+import { useQueryState } from 'nuqs';
+
+interface Profile {
+  git?: {
+    login: string;
+    provider: 'github' | 'gitlab';
+  };
+  id?: string;
+  image?: string;
+  name?: string;
+}
+
+interface PullRequestData {
+  mergedAt?: string;
+  isDraft?: boolean;
+  createdAt: string;
+  updatedAt?: string;
+  state: 'open' | 'closed' | 'merged';
+  url: string;
+  title: string;
+  repository: {
+    url: string;
+    nameWithOwner: string;
+  };
+  headRefName?: string;
+  baseRefName?: string;
+  id: string;
+}
+
+interface ProfileTabsProps {
+  profile: Profile | undefined;
+  isProfileLoading: boolean;
+  tab: string;
+  setTab: (value: string) => void;
+  featuredProjects: any[];
+  projectsWithGithubData: any[];
+}
+
+export function ProfileTabs({
+  profile,
+  isProfileLoading,
+  tab,
+  setTab,
+  featuredProjects,
+  projectsWithGithubData,
+}: ProfileTabsProps) {
+  return (
+    <>
+      {isProfileLoading && (
+        <Card className="rounded-none border-neutral-800 bg-neutral-900/50 p-0">
+          <CardContent className="p-4">
+            <div className="flex items-center justify-center py-8">
+              <div className="text-neutral-400">Loading contribution graph...</div>
+            </div>
+          </CardContent>
+        </Card>
+      )}
+      {!isProfileLoading && profile?.git && (
+        <ContributionGraph username={profile.git.login} provider={profile.git.provider} />
+      )}
+      <Tabs defaultValue={tab} onValueChange={setTab} className="w-full">
+        <TabsList className="grid w-full grid-cols-3 rounded-none border-neutral-800 bg-neutral-900/50">
+          <TabsTrigger value="projects" className="rounded-none">
+            Projects
+          </TabsTrigger>
+          <TabsTrigger value="contributions" className="rounded-none">
+            Contributions
+          </TabsTrigger>
+          <TabsTrigger value="collections" className="rounded-none">
+            Collections
+          </TabsTrigger>
+        </TabsList>
+
+        <TabsContent value="projects" className="mt-2">
+          {featuredProjects.length > 0 ? (
+            <div className="mb-8">
+              <div className="mb-4 flex items-center justify-between">
+                <h2 className="flex items-center text-xl font-semibold">
+                  <TrendingUp className="mr-2 h-5 w-5 text-orange-500" />
+                  Featured Projects
+                </h2>
+                <div className="flex space-x-2">
+                  <Button
+                    variant="ghost"
+                    size="sm"
+                    onClick={() => {
+                      const container = document.getElementById('featured-carousel');
+                      if (container) {
+                        container.scrollBy({ left: -300, behavior: 'smooth' });
+                      }
+                    }}
+                    className="h-8 w-8 p-0"
+                  >
+                    ←
+                  </Button>
+                  <Button
+                    variant="ghost"
+                    size="sm"
+                    onClick={() => {
+                      const container = document.getElementById('featured-carousel');
+                      if (container) {
+                        container.scrollBy({ left: 300, behavior: 'smooth' });
+                      }
+                    }}
+                    className="h-8 w-8 p-0"
+                  >
+                    →
+                  </Button>
+                </div>
+              </div>
+              <div
+                id="featured-carousel"
+                className="flex space-x-4 overflow-x-auto pb-4"
+                style={{
+                  scrollbarWidth: 'none',
+                  msOverflowStyle: 'none',
+                }}
+              >
+                {featuredProjects.length > 0 ? (
+                  featuredProjects?.map((project) => (
+                    <ProjectCard key={project.id} project={project} />
+                  ))
+                ) : (
+                  <div className="flex items-center justify-center">
+                    <p className="text-neutral-400">No featured projects</p>
+                  </div>
+                )}
+              </div>
+            </div>
+          ) : null}
+
+          <div>
+            <div className="space-y-4">
+              {projectsWithGithubData?.map((project) => (
+                <ProjectCard key={project.id} project={project} />
+              ))}
+            </div>
+          </div>
+        </TabsContent>
+
+        <TabsContent value="contributions" className="mt-2">
+          <UserPullRequests profile={profile as Profile} />
+        </TabsContent>
+
+        <TabsContent value="collections" className="mt-2">
+          <div className="py-12 text-center">
+            <div className="mb-4 text-neutral-400">
+              <Heart className="mx-auto mb-4 h-12 w-12 opacity-50" />
+              <p>Project collections coming soon...</p>
+            </div>
+          </div>
+        </TabsContent>
+      </Tabs>
+    </>
+  );
+}
+
+function UserPullRequests({ profile }: { profile: Profile }) {
+  const trpc = useTRPC();
+  const [selectedState, setSelectedState] = useQueryState('selectedState', {
+    defaultValue: 'all',
+  });
+  const [sortBy, setSortBy] = useQueryState('sortBy', {
+    defaultValue: 'recent',
+  });
+
+  const { data, isLoading, error } = useQuery(
+    trpc.profile.getUserPullRequests.queryOptions(
+      {
+        username: profile?.git?.login ?? '',
+        provider: profile?.git?.provider ?? 'github',
+        state: selectedState as 'all' | 'open' | 'closed' | 'merged',
+        limit: 100,
+      },
+      {
+        enabled: !!profile?.git?.login && !!profile?.git?.provider,
+      },
+    ),
+  );
+
+  if (!profile?.git?.login || !profile?.git?.provider) {
+    return (
+      <div className="py-12 text-center">
+        <div className="mb-4 text-neutral-400">
+          <Icons.github className="mx-auto mb-4 h-12 w-12 opacity-50" />
+          <p>No Git provider connected to this profile</p>
+        </div>
+      </div>
+    );
+  }
+
+  if (isLoading) {
+    return (
+      <div className="space-y-4">
+        {[...Array(3)].map((_, i) => (
+          <Card key={i} className="rounded-none border-neutral-800 bg-neutral-900/50 p-0">
+            <CardContent className="p-4">
+              <div className="flex items-start justify-between">
+                <div className="flex-1">
+                  <Skeleton className="mb-2 h-5 w-3/4 rounded-none" />
+                  <Skeleton className="mb-2 h-4 w-1/2 rounded-none" />
+                  <Skeleton className="h-4 w-1/4 rounded-none" />
+                </div>
+              </div>
+            </CardContent>
+          </Card>
+        ))}
+      </div>
+    );
+  }
+
+  if (error) {
+    return (
+      <div className="py-12 text-center">
+        <div className="mb-4 text-neutral-400">
+          <Icons.github className="mx-auto mb-4 h-12 w-12 opacity-50" />
+          <p>Failed to load pull requests</p>
+          <p className="mt-2 text-sm">{error.message}</p>
+        </div>
+      </div>
+    );
+  }
+
+  const pullRequests = data?.pullRequests || [];
+
+  const calculatePRScore = (pr: PullRequestData) => {
+    let score = 0;
+
+    if (pr.mergedAt) score += 20;
+    if (pr.isDraft) score -= 10;
+
+    const ageInDays = Math.floor(
+      (Date.now() - new Date(pr.createdAt).getTime()) / (1000 * 60 * 60 * 24),
+    );
+    if (ageInDays < 30) score += 5;
+
+    if (pr.mergedAt && pr.createdAt) {
+      const daysToMerge = Math.floor(
+        (new Date(pr.mergedAt).getTime() - new Date(pr.createdAt).getTime()) /
+          (1000 * 60 * 60 * 24),
+      );
+      if (daysToMerge < 3) score += 10;
+      else if (daysToMerge > 30) score -= 5;
+    }
+
+    return Math.round(score);
+  };
+
+  const sortedPullRequests = [...pullRequests].sort((a, b) => {
+    switch (sortBy) {
+      case 'impact':
+        return calculatePRScore(b as PullRequestData) - calculatePRScore(a as PullRequestData);
+      case 'discussion':
+        return (
+          new Date(b.updatedAt || b.createdAt).getTime() -
+          new Date(a.updatedAt || a.createdAt).getTime()
+        );
+      case 'fastest': {
+        const aTime = a.mergedAt
+          ? new Date(a.mergedAt).getTime() - new Date(a.createdAt).getTime()
+          : Infinity;
+        const bTime = b.mergedAt
+          ? new Date(b.mergedAt).getTime() - new Date(b.createdAt).getTime()
+          : Infinity;
+        return aTime - bTime;
+      }
+      default:
+        return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
+    }
+  });
+
+  const stats = {
+    open: pullRequests.filter((pr) => pr.state === 'open').length,
+    closed: pullRequests.filter((pr) => pr.state === 'closed' && !pr.mergedAt).length,
+    merged: pullRequests.filter((pr) => pr.mergedAt).length,
+    avgMergeTime: (() => {
+      const mergedPRs = pullRequests.filter((pr) => pr.mergedAt);
+      if (mergedPRs.length === 0) return 0;
+      const totalTime = mergedPRs.reduce((sum, pr) => {
+        return sum + (new Date(pr.mergedAt!).getTime() - new Date(pr.createdAt).getTime());
+      }, 0);
+      return Math.round(totalTime / mergedPRs.length / (1000 * 60 * 60 * 24));
+    })(),
+  };
+
+  return (
+    <div>
+      <div className="mb-6 grid grid-cols-2 gap-4 sm:grid-cols-4">
+        <Card className="rounded-none border-neutral-800 bg-neutral-900/50 p-0">
+          <CardContent className="p-4">
+            <div className="text-2xl font-bold">{stats.merged}</div>
+            <div className="text-xs text-neutral-400">Merged PRs</div>
+          </CardContent>
+        </Card>
+        <Card className="rounded-none border-neutral-800 bg-neutral-900/50 p-0">
+          <CardContent className="p-4">
+            <div className="text-2xl font-bold">{stats.open}</div>
+            <div className="text-xs text-neutral-400">Open PRs</div>
+          </CardContent>
+        </Card>
+        <Card className="rounded-none border-neutral-800 bg-neutral-900/50 p-0">
+          <CardContent className="p-4">
+            <div className="text-2xl font-bold">{stats.avgMergeTime}d</div>
+            <div className="text-xs text-neutral-400">Avg Merge Time</div>
+          </CardContent>
+        </Card>
+        <Card className="rounded-none border-neutral-800 bg-neutral-900/50 p-0">
+          <CardContent className="p-4">
+            <div className="text-2xl font-bold">{pullRequests.length}</div>
+            <div className="text-xs text-neutral-400">Total PRs</div>
+          </CardContent>
+        </Card>
+      </div>
+
+      <div className="mb-6 flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
+        <h2 className="text-xl font-semibold">Pull Requests</h2>
+        <div className="xs:flex-row xs:gap-2 flex w-full gap-2 sm:w-auto">
+          <Select value={sortBy} onValueChange={(value: string) => setSortBy(value)}>
+            <SelectTrigger className="xs:w-[140px] w-full rounded-none">
+              <SelectValue />
+            </SelectTrigger>
+            <SelectContent className="left-0 rounded-none">
+              <SelectItem value="recent">Most Recent</SelectItem>
+              <SelectItem value="impact">By Status</SelectItem>
+              <SelectItem value="discussion">Last Updated</SelectItem>
+              <SelectItem value="fastest">Fastest Merged</SelectItem>
+            </SelectContent>
+          </Select>
+          <Select value={selectedState} onValueChange={(value: string) => setSelectedState(value)}>
+            <SelectTrigger className="xs:w-[140px] w-full rounded-none">
+              <SelectValue />
+            </SelectTrigger>
+            <SelectContent className="left-0 rounded-none">
+              <SelectItem value="all">All</SelectItem>
+              <SelectItem value="open">Open</SelectItem>
+              <SelectItem value="merged">Merged</SelectItem>
+              <SelectItem value="closed">Closed</SelectItem>
+            </SelectContent>
+          </Select>
+        </div>
+      </div>
+
+      {sortedPullRequests.length === 0 ? (
+        <div className="py-12 text-center">
+          <div className="mb-4 text-neutral-400">
+            <GitFork className="mx-auto mb-4 h-12 w-12 opacity-50" />
+            <p>No {selectedState !== 'all' ? selectedState : ''} pull requests found</p>
+          </div>
+        </div>
+      ) : (
+        <div className="space-y-4">
+          {sortedPullRequests.map((pr) => {
+            // const score = calculatePRScore(pr);
+            const isMergedQuickly =
+              pr.mergedAt &&
+              new Date(pr.mergedAt).getTime() - new Date(pr.createdAt).getTime() <
+                3 * 24 * 60 * 60 * 1000;
+
+            return (
+              <Card
+                key={pr.id}
+                className={cn(
+                  'rounded-none border-neutral-800 bg-neutral-900/50 p-0 transition-colors hover:bg-neutral-900/70',
+                  isMergedQuickly && 'border-green-900/50',
+                )}
+              >
+                <CardContent className="p-4">
+                  <div className="flex flex-col gap-2 sm:flex-row sm:items-start sm:justify-between">
+                    <div className="flex-1">
+                      <div className="mb-2 flex flex-wrap items-center gap-2">
+                        <Link
+                          href={pr.url}
+                          target="_blank"
+                          className="hover:text-primary text-base font-medium"
+                        >
+                          {pr.title}
+                        </Link>
+                        <Badge
+                          variant={
+                            pr.state === 'open' ? 'default' : pr.mergedAt ? 'secondary' : 'outline'
+                          }
+                          className={cn(
+                            'rounded-none text-xs',
+                            pr.state === 'open' &&
+                              'border-green-800 bg-green-900/20 text-green-400',
+                            pr.mergedAt && 'border-purple-800 bg-purple-900/20 text-purple-400',
+                            pr.state === 'closed' &&
+                              !pr.mergedAt &&
+                              'border-red-800 bg-red-900/20 text-red-400',
+                          )}
+                        >
+                          {pr.mergedAt ? 'Merged' : pr.state}
+                        </Badge>
+                        {pr.isDraft && (
+                          <Badge variant="outline" className="rounded-none text-xs">
+                            Draft
+                          </Badge>
+                        )}
+                        {isMergedQuickly && (
+                          <Badge className="rounded-none border-green-800 bg-green-900/20 text-xs text-green-400">
+                            <Clock className="mr-1 h-3 w-3" />
+                            Quick Merge
+                          </Badge>
+                        )}
+                      </div>
+                      <div className="mb-1 flex flex-col sm:flex-row sm:items-center sm:justify-between">
+                        <div className="flex items-center text-sm text-neutral-400">
+                          <Link
+                            href={pr.repository.url}
+                            target="_blank"
+                            className="hover:text-neutral-200"
+                          >
+                            {pr.repository.nameWithOwner}
+                          </Link>
+                        </div>
+                        <div className="mt-1 flex flex-wrap items-center gap-4 text-xs text-neutral-400 sm:mt-0">
+                          <span className="flex items-center gap-1">
+                            <Icons.clock className="h-3 w-3" />
+                            {new Date(pr.createdAt).toLocaleDateString()}
+                          </span>
+                          {pr.mergedAt && (
+                            <span className="flex items-center gap-1">
+                              <Icons.merge className="h-3 w-3" />
+                              {new Date(pr.mergedAt).toLocaleDateString()}
+                            </span>
+                          )}
+                        </div>
+                      </div>
+                      {pr.headRefName && (
+                        <div className="mt-2 flex items-center justify-between gap-2 text-xs text-neutral-500">
+                          <div className="flex items-center gap-2">
+                            <code className="rounded-none bg-neutral-800 px-1.5 py-0.5">
+                              {pr.headRefName}
+                            </code>
+                            <span>→</span>
+                            <code className="rounded-none bg-neutral-800 px-1.5 py-0.5">
+                              {pr.baseRefName}
+                            </code>
+                          </div>
+                          <Button variant="ghost" size="sm" asChild>
+                            <Link href={pr.url} target="_blank">
+                              <ExternalLink className="h-4 w-4" />
+                            </Link>
+                          </Button>
+                        </div>
+                      )}
+                      {!pr.headRefName && (
+                        <div className="mt-2 flex justify-end">
+                          <Button variant="ghost" size="sm" asChild>
+                            <Link href={pr.url} target="_blank">
+                              <ExternalLink className="h-4 w-4" />
+                            </Link>
+                          </Button>
+                        </div>
+                      )}
+                    </div>
+                  </div>
+                </CardContent>
+              </Card>
+            );
+          })}
+        </div>
+      )}
+    </div>
+  );
+}
